fix(shell-api): map `fields` to `projection` in `findAndModify` MONGOSH-1820 (#2100)

+++ packages/shell-api/src/collection.ts
@@ -482,8 +482,19 @@ export default class Collection extends ShellApiWithMongoClass {
       FindAndModifyMethodShellOptions,
       'query' | 'update'
     > = { ...options };
+    if (
+      reducedOptions.projection !== undefined &&
+      reducedOptions.fields !== undefined
+    ) {
+      throw new MongoshInvalidInputError(
+        'Cannot specify both .fields and .projection for findAndModify()',
+        CommonErrors.InvalidArgument
+      );
+    }
+    reducedOptions.projection ??= reducedOptions.fields;
     delete (reducedOptions as any).query;
     delete (reducedOptions as any).update;
+    delete (reducedOptions as any).fields;
     if (options.remove) {
       return this.findOneAndDelete(options.query, reducedOptions);
     }
